requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
requests==2.31.0

# LLM and AI integration
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
python-dotenv==1.0.0

# Database integration
supabase==2.0.2
postgrest==0.13.2

# Data validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Optional: For HTML templates
jinja2==3.1.2 